+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## develop branch
 
-Nothing yet
+### New
+
+- CodeCaller value builder (moved from ganbarodigital/php-reflection)
 
 ## 1.0.0 - Sat Jun 27 2015