+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## develop branch
 
+Nothing yet.
+
+## 1.1.0 - Sat Jul 4 2015
+
 ### New
 
 - CodeCaller value builder (moved from ganbarodigital/php-reflection)